third_party/WebKit/LayoutTests/http/tests/feature-policy/vibrate-disabled-expected.txt

CONSOLE ERROR: line 9: Blocked call to navigator.vibrate because user hasn't tapped on the frame or any embedded frame yet: https://www.chromestatus.com/feature/5644273861001216.
 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
This is a testharness.js-based test.
FAIL No iframe may call navigator.vibrate when disabled. assert_false: expected false got true
Harness: the test ran to completion.


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
This is a testharness.js-based test.
PASS No iframe may call navigator.vibrate when disabled. 
Harness: the test ran to completion.
